+++ src/renderer/components/caption-studio/JsonPreview.vue
@@ -58,7 +58,7 @@ import { EventBus } from '@/renderer/class/EventBus';
 import VJsoneditor from 'v-jsoneditor';
 import { ipcRenderer, path } from 'electron';
 import { mapState } from 'vuex';
-import { EVENTS } from '../../contents';
+import { EVENTS } from '@/constants';
 
 const fs = require('fs');
 
@@ -140,12 +140,11 @@ export default {
   },
   methods: {
     /**
-     *
+     * Handles JSON Editor changes
      */
     onEdit($event) {
       this.checkErrors($event, this.origin);
       if (this.jsonErrors) {
-        console.log('errors?', this.jsonErrors);
         return;
       }
       EventBus.$emit('json_update', $event, this.origin);
@@ -156,6 +155,9 @@ export default {
     onSave(event, filePath, force = false) {
 
       if (!filePath) {
+        if (!this.captionLocation) {
+          return;
+        }
         filePath = this.captionLocation;
       }
 
@@ -170,7 +172,6 @@ export default {
           throw err;
         }
         this.$store.dispatch('setIsUnsavedChanges', { isUnsavedChanges: false });
-        console.log('JSON data is saved.');
       });
     },
     /**
@@ -188,14 +189,14 @@ export default {
       this.dialog = true;
     },
     /**
-     *
+     * When a caption updated in another component, requests the new caption data from CaptionManager
      */
     onUpdate(data, $origin) {
       //Pass the origin of the original component on through in this call, since that is the origin that matters
       EventBus.$emit('caption_get', $origin);
     },
     /**
-     *
+     * Handler for any JSON editor events, filters out any that aren't the 'focus' event.
      */
     onEvent(node, event) {
       if (event.type !== 'focus') {
@@ -221,7 +222,7 @@ export default {
 
     },
     /**
-     *
+     * Handles the "caption_changed" event and updates data to match the new caption
      */
     onCaptionChange({
       index,
@@ -233,7 +234,7 @@ export default {
       this.currentIndex = index;
     },
     /**
-     *
+     * Updates current JSON data when CaptionManager emits the new caption data
      */
     update(data, $origin) {
       this.checkErrors(data, $origin);
@@ -250,7 +251,7 @@ export default {
 
     },
     /**
-     *
+     * Creates a map of file names to their files, use to help with switching active files
      */
     createFileNameMap($event) {
       if (!Array.isArray($event)) {
@@ -262,7 +263,7 @@ export default {
       });
     },
     /**
-     *
+     * Strips out any invalid captions in the json data to avoid caption renderer errors
      */
     cleanData(data) {
       const key = Object.keys(data);
@@ -292,7 +293,7 @@ export default {
       return output;
     },
     /**
-     *
+     * Creates a blob out of the current json data for user download
      */
     createBlob() {
       this.blob = URL.createObjectURL(
@@ -302,7 +303,7 @@ export default {
       );
     },
     /**
-     *
+     * Resets/clears all caption data
      */
     reset() {
       EventBus.$emit('caption_reset');
@@ -316,7 +317,13 @@ export default {
       const errors = {};
       Object.keys(json).forEach(key => {
         errors[key] = [];
+
         const file = json[key];
+
+        if (!Array.isArray(file)) {
+          return;
+        }
+
         file.forEach((caption, index) => {
           if (caption.edited || $origin === this.origin) {
             if (!caption.content || !caption.content.trim()) {
@@ -338,11 +345,10 @@ export default {
           delete errors[key];
         }
       });
-
       return errors;
     },
     /**
-     *
+     * Checks data for error and emits them out for any other components
      */
     checkErrors(data, $origin) {
       this.jsonErrors = false;