+++ testing/unit/renderer/specs/caption-studio/TimeStampInput.spec.js
@@ -1,0 +1,10 @@
+import { createVue } from '../../utils/VueUtils';
+import TimeStampInput from '@/renderer/components/caption-studio/TimeStampInput.vue';
+
+describe('TimeStampInput.vue', () => {
+
+  it('should mount successfully', () => {
+    const wrapper = createVue(TimeStampInput);
+  });
+
+});