+++ testing/unit/renderer/specs/store/PersistentState.spec.js
@@ -1,4 +1,5 @@
 import { createVue } from '../../utils/VueUtils';
+import { mount } from '@vue/test-utils';
 import LandingPage from '@/renderer/components/pages/LandingPage.vue';
 
 describe('PersistentState.js', () => {
@@ -31,5 +32,25 @@ describe('PersistentState.js', () => {
     wrapper = createVue(LandingPage);
 
     expect(wrapper.vm.$store.state.gamePreview.previewURL).to.equal('this/is/a/test');
+  });
+
+  it('should persist the audio file location', () => {
+    let wrapper = createVue(LandingPage);
+    wrapper.vm.$store.commit('audioLocation', { audioLocation: 'this/is/also/a/test' });
+
+    wrapper.destroy();
+    wrapper = createVue(LandingPage);
+
+    expect(wrapper.vm.$store.state.captionInfo.audioLocation).to.equal('this/is/also/a/test');
+  });
+
+  it('should persist the caption file location', () => {
+    let wrapper = createVue(LandingPage);
+    wrapper.vm.$store.commit('captionLocation', { captionLocation: 'this/is/another/test' });
+
+    wrapper.destroy();
+    wrapper = createVue(LandingPage);
+
+    expect(wrapper.vm.$store.state.captionInfo.captionLocation).to.equal('this/is/another/test');
   });
 });