+++ testing/unit/renderer/utils/VueUtils.js
@@ -1,6 +1,6 @@
 import { mount, createLocalVue } from '@vue/test-utils';
 import Vuex from 'vuex';
-// import Vuetify from 'vuetify';
+
 import persistentState from '@/renderer/store/storage/PersistentState';
 import projectInfo from '@/renderer/store/modules/ProjectInfo';
 import gamePreview from '@/renderer/store/modules/GamePreview';
@@ -9,7 +9,6 @@ import captionInfo from '@/renderer/store/modules/CaptionInfo';
 const localVue = createLocalVue();
 
 localVue.use(Vuex);
-// localVue.use(Vuetify);
 
 export const createVue = (component, options = {}) => {
   return mount(component, {