+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0] - UNRELEASED
+### Added
+- Unit tests added for caption studio components and classes
 ### Changed
 - Caption Studio now reads audio files directly from project root, or user specified audio directory
 - Caption studio now allows for saving/opening caption JSON files directly from projects